chore: Remove forceful cleanup of all containers
A while back it was introduced to cleanup the containers due to caching
issue. This also cleans up your own containers and should not be ran
locally whatsoever as it forcefully removes all containers.

This is just a removal for now. It is better to have to run the command
twice than have to deal with the aftermath of all your own containers
suddenly being gone.